+++ core/components/instafeed/lexicon/en/setting.inc.php
@@ -13,3 +13,5 @@ $_lang['setting_instafeed.remote'] = 'Remote Server';
 $_lang['setting_instafeed.remote_desc'] = 'Use Remote Server?';
 $_lang['setting_instafeed.remote_key'] = 'Remote Server Key';
 $_lang['setting_instafeed.remote_key_desc'] = 'Remote Server Auth Key.';
+$_lang['setting_instafeed.published'] = 'Publish new Posts';
+$_lang['setting_instafeed.published_desc'] = 'Publish new posts automaticaly?';