+++ core/components/instafeed/docs/changelog.md
@@ -1,5 +1,12 @@
 Changelog for instaFeed
 
+instaFeed 1.0.2
+---------------------------------
++ Clear cache on import
++ Add published system setting
++ Add likes & comments property
++ update existing properties of an post
+
 instaFeed 1.0.1
 ---------------------------------
 + Add remote Service